Create README.md
To use this program, run the python file through the terminal with the following command:
python musicPlayer.pyw

Usage:
To add music to the playlist, go to File->Open and select the music you'd like to add or use the add button at the bottom of the playlist. 
After you've added music to the playlist, highlight your desired track and use the play button to start it.